Moved SourceType loading to SourceTypeService
